requirements.txt
Flask==3.0.0
scikit-learn==1.3.1
pandas==2.1.1
numpy==1.26.0
joblib==1.3.2
urllib3
whois
tldextract
python-Levenshtein
